enhancement: use deploying wallet at the end for voting contests
