+++ packages/react-app-revamp/hooks/useDeployContest/index.ts
@@ -1,25 +1,22 @@
 import { toastError, toastLoading, toastSuccess } from "@components/UI/Toast";
-import { chains, config } from "@config/wagmi";
+import { config } from "@config/wagmi";
 import DeployedContestContract from "@contracts/bytecodeAndAbi//Contest.sol/Contest.json";
-import { MAX_ROWS } from "@helpers/csvConstants";
 import { isSupabaseConfigured } from "@helpers/database";
 import { getEthersSigner } from "@helpers/ethers";
-import getContestContractVersion from "@helpers/getContestContractVersion";
-import { isR2Configured } from "@helpers/r2";
-import useV3ContestsIndex, { ContestValues } from "@hooks/useContestsIndexV3";
 import useEmailSignup from "@hooks/useEmailSignup";
 import { useError } from "@hooks/useError";
-import { readContract } from "@wagmi/core";
 import { differenceInSeconds, getUnixTime } from "date-fns";
-import { ContractFactory, formatUnits, JsonRpcSigner } from "ethers";
-import { loadFileFromBucket, saveFileToBucket } from "lib/buckets";
-import { Recipient } from "lib/merkletree/generateMerkleTree";
-import { canUploadLargeAllowlist } from "lib/vip";
-import { Abi, parseEther } from "viem";
+import { ContractFactory, JsonRpcSigner } from "ethers";
+import { parseEther } from "viem";
 import { useAccount } from "wagmi";
+import { saveFilesToBucket } from "./buckets";
+import { isSortingEnabled } from "./contracts";
+import { checkForSpoofing, getJkLabsSplitDestinationAddress, indexContest } from "./database";
+import { createMetadataFieldsSchema } from "./helpers";
 import { useDeployContestStore } from "./store";
-import { PriceCurveType, SplitFeeDestinationType, SubmissionMerkle, VoteType, VotingMerkle } from "./types";
-import { EntryPreviewConfig, MetadataField } from "./slices/contestMetadataSlice";
+import { PriceCurveType, SplitFeeDestinationType, VoteType } from "./types";
+import { useSubmissionMerkle } from "@components/_pages/Create/hooks/useSubmissionMerkle";
+import { ContestType } from "@components/_pages/Create/types";
 
 export const MAX_SUBMISSIONS_LIMIT = 1000;
 export const JK_LABS_SPLIT_DESTINATION_DEFAULT = "0xDc652C746A8F85e18Ce632d97c6118e8a52fa738";
@@ -27,7 +24,6 @@ export const JK_LABS_SPLIT_DESTINATION_DEFAULT = "0xDc652C746A8F85e18Ce632d97c61
 const EMPTY_ROOT = "0x0000000000000000000000000000000000000000000000000000000000000000";
 
 export function useDeployContest() {
-  const { indexContestV3 } = useV3ContestsIndex();
   const { subscribeUser, checkIfEmailExists } = useEmailSignup();
   const {
     title,
@@ -36,8 +32,6 @@ export function useDeployContest() {
     submissionOpen,
     votingOpen,
     votingClose,
-    votingMerkle: votingMerkleData,
-    submissionMerkle: submissionMerkleData,
     customization,
     advancedOptions,
     setDeployContestData,
@@ -50,34 +44,75 @@ export function useDeployContest() {
     setIsLoading,
     setIsSuccess,
   } = useDeployContestStore(state => state);
-  const { error, handleError } = useError();
+  const { handleError } = useError();
   const { address, chain } = useAccount();
+  const { processCreatorAllowlist } = useSubmissionMerkle();
 
-  async function deployContest() {
+  async function prepareForDeployment() {
+    if (contestType === ContestType.VotingContest) {
+      try {
+        await processCreatorAllowlist(address);
+      } catch (error) {
+        handleError(error, "Something went wrong while processing creator allowlist.");
+        throw error;
+      }
+    }
+
     let signer: JsonRpcSigner;
-
     try {
       signer = await getEthersSigner(config, { chainId: chain?.id });
     } catch (error: any) {
       handleError(error, "Please try reconnecting your wallet.");
-      return;
+      throw error;
     }
 
-    const isSpoofingDetected = await checkForSpoofing(signer?.address);
+    const currentState = useDeployContestStore.getState();
+    const currentVotingMerkleData = currentState.votingMerkle;
+    const currentSubmissionMerkleData = currentState.submissionMerkle;
+
+    const isSpoofingDetected = await checkForSpoofing(
+      signer?.address,
+      currentVotingMerkleData,
+      currentSubmissionMerkleData,
+    );
 
     if (isSpoofingDetected) {
       toastError({
         message: "Spoofing detected! None shall pass.",
       });
+      throw new Error("Spoofing detected");
+    }
+
+    return {
+      signer,
+      votingMerkleData: currentVotingMerkleData,
+      submissionMerkleData: currentSubmissionMerkleData,
+    };
+  }
+
+  async function deployContest() {
+    setIsLoading(true);
+    toastLoading({
+      message: "contest is deploying...",
+    });
+
+    let preparationResult;
+    try {
+      preparationResult = await prepareForDeployment();
+    } catch (error) {
       setIsLoading(false);
       return;
     }
 
-    setIsLoading(true);
+    const {
+      signer,
+      votingMerkleData: currentVotingMerkleData,
+      submissionMerkleData: currentSubmissionMerkleData,
+    } = preparationResult;
 
-    toastLoading({
-      message: "contest is deploying...",
-    });
+    console.log("currentVotingMerkleData", currentVotingMerkleData);
+    console.log("currentSubmissionMerkleData", currentSubmissionMerkleData);
+
     try {
       const factoryCreateContest = new ContractFactory(
         DeployedContestContract.abi,
@@ -93,8 +128,8 @@ export function useDeployContest() {
         imageUrl: prompt.imageUrl ?? "",
       }).toString();
 
-      const votingMerkle = votingMerkleData.prefilled || votingMerkleData.csv;
-      const submissionMerkle = submissionMerkleData;
+      const votingMerkle = currentVotingMerkleData.prefilled || currentVotingMerkleData.csv;
+      const submissionMerkle = currentSubmissionMerkleData;
       const { type: chargeType, percentageToCreator } = charge;
       const { merkleRoot: submissionMerkleRoot = EMPTY_ROOT } = submissionMerkle || {};
 
@@ -147,7 +182,7 @@ export function useDeployContest() {
             ? jkLabsSplitDestination || JK_LABS_SPLIT_DESTINATION_DEFAULT
             : charge.splitFeeDestination.address,
         jkLabsSplitDestination: jkLabsSplitDestination || JK_LABS_SPLIT_DESTINATION_DEFAULT,
-        metadataFieldsSchema: createMetadataFieldsSchema(metadataFields),
+        metadataFieldsSchema: createMetadataFieldsSchema(metadataFields, entryPreviewConfig),
       };
 
       const contractContest = await factoryCreateContest.deploy(
@@ -175,7 +210,7 @@ export function useDeployContest() {
 
       let votingReqDatabaseEntry = null;
 
-      if (votingMerkleData.prefilled) {
+      if (currentVotingMerkleData.prefilled) {
         votingReqDatabaseEntry = {
           type: votingRequirements.type,
           tokenAddress: votingRequirements.tokenAddress,
@@ -205,8 +240,24 @@ export function useDeployContest() {
       };
 
       await subscribeToEmail(emailSubscriptionAddress);
-      await saveFilesToBucket(votingMerkle, submissionMerkle);
-      await indexContest(contestData, votingMerkle, submissionMerkle);
+
+      try {
+        await saveFilesToBucket(votingMerkle, submissionMerkle);
+      } catch (e) {
+        handleError(e, "Something went wrong while saving files to bucket.");
+        setIsLoading(false);
+        throw e;
+      }
+
+      try {
+        await indexContest(contestData, votingMerkle, submissionMerkle);
+      } catch (e) {
+        setIsLoading(false);
+        toastError({
+          message: "contest deployment failed to index in db",
+        });
+        throw e;
+      }
 
       toastSuccess({
         message: "contest has been deployed!",
@@ -216,40 +267,6 @@ export function useDeployContest() {
     } catch (e) {
       handleError(e, "Something went wrong and the contest couldn't be deployed.");
       setIsLoading(false);
-    }
-  }
-
-  async function saveFilesToBucket(votingMerkle: VotingMerkle | null, submissionMerkle: SubmissionMerkle | null) {
-    if (!isR2Configured) {
-      throw new Error("R2 is not configured");
-    }
-
-    const tasks: Promise<void>[] = [];
-
-    if (votingMerkle && !(await checkExistingFileInBucket(votingMerkle.merkleRoot))) {
-      tasks.push(
-        saveFileToBucket({
-          fileId: votingMerkle.merkleRoot,
-          content: formatRecipients(votingMerkle.voters),
-        }),
-      );
-    }
-
-    if (submissionMerkle && !(await checkExistingFileInBucket(submissionMerkle.merkleRoot))) {
-      tasks.push(
-        saveFileToBucket({
-          fileId: submissionMerkle.merkleRoot,
-          content: formatRecipients(submissionMerkle.submitters),
-        }),
-      );
-    }
-
-    try {
-      await Promise.all(tasks);
-    } catch (e) {
-      handleError(e, "Something went wrong while saving files to bucket.");
-      setIsLoading(false);
-      throw e;
     }
   }
 
@@ -271,214 +288,8 @@ export function useDeployContest() {
     await subscribeUser(emailAddress, address, false);
   }
 
-  async function checkExistingFileInBucket(fileId: string): Promise<boolean> {
-    try {
-      const existingData = await loadFileFromBucket({ fileId });
-      return !!(existingData && existingData.length > 0);
-    } catch (e) {
-      return false;
-    }
-  }
-
-  async function indexContest(
-    contestData: ContestValues,
-    votingMerkle: VotingMerkle | null,
-    submissionMerkle: SubmissionMerkle | null,
-  ) {
-    const participantsWorker = new Worker(new URL("/workers/indexContestParticipants", import.meta.url));
-
-    try {
-      if (!isSupabaseConfigured) {
-        throw new Error("Supabase is not configured");
-      }
-
-      const tasks = [];
-
-      tasks.push(indexContestV3(contestData));
-
-      const workerData = {
-        contestData,
-        votingMerkle,
-        submissionMerkle,
-      };
-
-      const workerTask = new Promise<void>((resolve, reject) => {
-        participantsWorker.onmessage = event => {
-          if (event.data.success) {
-            resolve();
-          } else {
-            reject(new Error(event.data.error));
-          }
-        };
-
-        participantsWorker.onerror = error => {
-          setIsLoading(false);
-          toastError({
-            message: "contest deployment failed to index in db",
-          });
-          reject(error);
-        };
-
-        participantsWorker.postMessage(workerData);
-      });
-
-      tasks.push(workerTask);
-
-      await Promise.all(tasks);
-    } catch (e: any) {
-      setIsLoading(false);
-      toastError({
-        message: "contest deployment failed to index in db",
-      });
-      throw e;
-    } finally {
-      participantsWorker.terminate();
-    }
-  }
-
-  async function checkForSpoofing(address: string) {
-    const votingMerkle = votingMerkleData.prefilled || votingMerkleData.csv;
-    const submissionMerkle = submissionMerkleData;
-
-    const exceedsVotingMaxRows = votingMerkle && votingMerkle.voters.length > MAX_ROWS;
-    const exceedsSubmissionMaxRows = submissionMerkle && submissionMerkle.submitters.length > MAX_ROWS;
-
-    let isVotingAllowListed = false;
-    let isSubmissionAllowListed = false;
-
-    if (exceedsVotingMaxRows) {
-      isVotingAllowListed = await canUploadLargeAllowlist(address, votingMerkle.voters.length);
-      if (!isVotingAllowListed) {
-        return true;
-      }
-    }
-
-    if (exceedsSubmissionMaxRows) {
-      isSubmissionAllowListed = await canUploadLargeAllowlist(address, submissionMerkle.submitters.length);
-      if (!isSubmissionAllowListed) {
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  async function isSortingEnabled(address: string, chainId: number) {
-    try {
-      const { abi } = await getContestContractVersion(address as `0x${string}`, chainId);
-
-      if (!abi) {
-        console.error("ABI not found");
-        return false;
-      }
-
-      const contractConfig = {
-        address: address as `0x${string}`,
-        abi: abi as Abi,
-        chainId: chainId,
-      };
-
-      const result = (await readContract(config, {
-        ...contractConfig,
-        functionName: "sortingEnabled",
-      })) as number;
-
-      return Number(result) === 1;
-    } catch (error) {
-      console.error("error in isSortingEnabled:", error);
-      return false;
-    }
-  }
-
-  async function getJkLabsSplitDestinationAddress(
-    chainId: number,
-    chargeType: { costToPropose: number; costToVote: number },
-  ): Promise<string> {
-    const chain = chains.find(c => c.id === chainId);
-
-    // check if costToVote is 0 ( this means no monetization )
-    if (chargeType.costToVote === 0) {
-      return "";
-    }
-
-    if (!chain) {
-      throw new Error(`Chain with id ${chainId} not found`);
-    }
-
-    if (!isSupabaseConfigured) {
-      throw new Error("Supabase is not configured");
-    }
-
-    const config = await import("@config/supabase");
-    const supabase = config.supabase;
-
-    const chainName = chain.name;
-
-    const { data, error } = await supabase
-      .from("chain_params")
-      .select("jk_labs_split_destination")
-      .eq("network_name", chainName.toLowerCase())
-      .single();
-
-    if (error) {
-      throw new Error(`Error fetching data: ${error.message}`);
-    }
-
-    if (!data) {
-      throw new Error(`No data found for chain ${chainName}`);
-    }
-
-    return data.jk_labs_split_destination;
-  }
-
-  function createMetadataFieldsSchema(metadataFields: MetadataField[]): string {
-    // start with an object that has a 'string' property initialized with the entry preview prompt
-    const initialSchema: Record<string, string | string[]> = {
-      string: getEntryPreviewPrompt(entryPreviewConfig),
-    };
-
-    const schema = metadataFields
-      .filter(field => field.prompt.trim() !== "")
-      .reduce<Record<string, string | string[]>>((acc, field) => {
-        const metadataType = field.metadataType;
-        const prompt = field.prompt.trim();
-
-        if (acc[metadataType]) {
-          if (Array.isArray(acc[metadataType])) {
-            (acc[metadataType] as string[]).push(prompt);
-          } else {
-            acc[metadataType] = [acc[metadataType] as string, prompt];
-          }
-        } else {
-          acc[metadataType] = prompt;
-        }
-
-        return acc;
-      }, initialSchema);
-
-    // ensure 'string' is always an array
-    if (!Array.isArray(schema.string)) {
-      schema.string = [schema.string];
-    }
-
-    return JSON.stringify(schema);
-  }
-
-  function getEntryPreviewPrompt(config: EntryPreviewConfig): string {
-    const { preview, isAdditionalDescriptionEnabled } = config;
-    const descriptionSuffix = isAdditionalDescriptionEnabled ? "_DESCRIPTION_ENABLED" : "_DESCRIPTION_NOT_ENABLED";
-    return `${preview}${descriptionSuffix}`;
-  }
-
-  // Helper function to format recipients (either voters or submitters)
-  function formatRecipients(recipients: Recipient[]): Recipient[] {
-    return recipients.map(recipient => ({
-      ...recipient,
-      numVotes: formatUnits(recipient.numVotes, 18),
-    }));
-  }
-
   return {
     deployContest,
+    prepareForDeployment,
   };
 }
